Update README.md
This change is to clarify the description which was rather misleading. It implied that the keyboard driven interface was the the **only** one available. Of course the app has a menu interface as well. I missed out on this excellent project for some time because I thought it would have a steep learning curve and be difficult to use.